requirements.txt
Django
djangorestframework
django-braces
channels
requests
django-crispy-forms
redisboard
hypercorn



